pyproject.toml
[project]
name = "react-resume-uv"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.5.0",
    "firebase-admin>=6.6.0",
    "google-cloud-firestore>=2.20.0",
    "streamlit>=1.42.0",
    "streamlit-authenticator>=0.4.1",
]
